+++ BNS360.Apis/Controllers/FeedbackController.cs
@@ -37,6 +37,10 @@ namespace BNS360.Apis.Controllers
                 return BadRequest(new ApiResponse(400, "Invalid user"));
             }
             var user = await _userManager.FindByEmailAsync(email);
+            if (user == null)
+            {
+                return Unauthorized(new ApiResponse(401, "Invalid user"));
+            }
             var result = await _feedbackRepository.AddFeedbackForBusiness(user.Id, businessId, feedback, rating);
             if (result.StatusCode == 400)
             {
@@ -58,6 +62,10 @@ namespace BNS360.Apis.Controllers
                 return BadRequest(new ApiResponse(400, "Invalid user"));
             }
             var user = await _userManager.FindByEmailAsync(email);
+            if (user == null)
+            {
+                return Unauthorized(new ApiResponse(401, "Invalid user"));
+            }
             var result = await _feedbackRepository.AddFeedbackForCraftsMen(user.Id, craftsMenId, feedback, rating);
             if (result.StatusCode == 400)
             {
@@ -79,6 +87,10 @@ namespace BNS360.Apis.Controllers
                 return BadRequest(new ApiResponse(400, "Invalid user"));
             }
             var user = await _userManager.FindByEmailAsync(email);
+            if (user == null)
+            {
+                return Unauthorized(new ApiResponse(401, "Invalid user"));
+            }
             var result = await _feedbackRepository.DeleteBusinessFeedback(user.Id, feedbackId);
             if (result.StatusCode == 400)
             {
@@ -100,6 +112,10 @@ namespace BNS360.Apis.Controllers
                 return BadRequest(new ApiResponse(400, "Invalid user"));
             }
             var user = await _userManager.FindByEmailAsync(email);
+            if (user == null)
+            {
+                return Unauthorized(new ApiResponse(401, "Invalid user"));
+            }
             var result = await _feedbackRepository.DeleteCraftsMenFeedback(user.Id, feedbackId);
             if (result.StatusCode == 400)
             {
