+++ BNS360.Apis/Controllers/PropertyController.cs
@@ -40,6 +40,10 @@ namespace BNS360.Apis.Controllers
                 return BadRequest(new ApiResponse(400, "Invalid user"));
             }
             var user = await _userManager.FindByEmailAsync(email);
+            if (user == null)
+            {
+                return Unauthorized(new ApiResponse(401, "Invalid user"));
+            }
             model.userId = user.Id;
             var result = await _propertyRepository.AddProperty(model);
             if (result.StatusCode == 400)
@@ -58,6 +62,10 @@ namespace BNS360.Apis.Controllers
                 return BadRequest(new ApiResponse(400, "Invalid user"));
             }
             var user = await _userManager.FindByEmailAsync(email);
+            if (user == null)
+            {
+                return Unauthorized(new ApiResponse(401, "Invalid user"));
+            }
             var result = await _propertyRepository.DeleteProperty(propertyId, user.Id);
             if (result.StatusCode == 400)
             {
@@ -99,6 +107,10 @@ namespace BNS360.Apis.Controllers
                 return BadRequest(new ApiResponse(400, "Invalid user"));
             }
             var user = await _userManager.FindByEmailAsync(email);
+            if (user == null)
+            {
+                return Unauthorized(new ApiResponse(401, "Invalid user"));
+            }
             model.userId = user.Id;
             var result = await _propertyRepository.UpdateProperty(PropertyId, model);
             if (result.StatusCode == 400)
